distanceの型をfloatからdoubleに変更

+++ tourism-app/Course.h
@@ -12,7 +12,7 @@
 @interface Course : NSObject {
     int courseid;
     NSString *course_name;
-    float distance;
+    double distance;
     int steps;
     int time;
     int male_calories;
@@ -22,7 +22,7 @@
 
 @property(nonatomic,assign)int courseid;
 @property(nonatomic,retain)NSString *course_name;
-@property(nonatomic,assign)float distance;
+@property(nonatomic,assign)double distance;
 @property(nonatomic,assign)int steps;
 @property(nonatomic,assign)int time;
 @property(nonatomic,assign)int male_calories;
